Add MetadaType crawl for attributes.
In instances where a class has a MetadataType defined, attributes may not be directly on the class, and the standard .GetCustomAttribute() methods will not locate attributes on the MetadataType. Added extension function to check for MetadataType and crawl it for attributes. This will also require passing in the original type, as it needs be checked for the MetadataTypeAttribute.